rename the floders name

+++ oss/oss.py
@@ -16,21 +16,22 @@ class OssSDK():
         self.bucket = oss2.Bucket(self.auth, self.endpoint, bucket)
         pass
 
-    def test(self):
+    def clear(self):
 
         print "test"
         #self.put('1.mp4','1.mp4')
         #print(self.bucket.sign_url('GET', 'RadioInterfaceTBox.pb.cc', 60))
         for object_info in oss2.ObjectIterator(self.bucket):
-           self.bucket.delete_object(object_info.key)
+            if object_info.key != 'NotFind.jpg':
+                self.bucket.delete_object(object_info.key)
         #result = self.bucket.get_object('RadioInterfaceTBox.pb.cc')
         #print dir(result)
 
     def percentage(consumed_bytes, total_bytes):
-        pass
-        # if total_bytes:
-        #     rate = int(100 * (float(consumed_bytes) / float(total_bytes)))
-        #     print '\r{0}% '.format(rate)
+        if total_bytes:
+            rate = int(100 * (float(consumed_bytes) / float(total_bytes)))
+            if rate == 100:
+                print '\r{0}% '.format(rate)
 
     def put_local(self,key,localfile,callback=percentage):
         return self.bucket.put_object_from_file(key,localfile,progress_callback=callback)
@@ -48,10 +49,16 @@ class OssSDK():
     def put_url_auto_name(self,url,callback=percentage):
         key = str(uuid.uuid1()) + '.jpg'
         if self.put_url(key,url,callback=callback):
+            print self.url + "/" + key
             return self.url + "/" + key
         else:
             return url
 
+    def put_content_auto_name(self,content,callback=percentage):
+        key = str(uuid.uuid1()) + '.jpg'
+        self.put_data(key,content,callback=callback)
+        print self.url + "/" + key
+        return self.url + "/" + key
 
     def put_data(self,key,data,callback=percentage):
         return  self.bucket.put_object(key,data,progress_callback=callback)
